+++ src/translation/screen_translation.rs
@@ -20,28 +20,28 @@ impl ScreenTranslation {
     };
   }
 
-  pub fn get_logical_position_x(&self, screen_x: usize) -> Option<usize> {
+  pub fn get_logical_position_x(&self, screen_x: usize) -> Option<i64> {
     if !self.in_screen_bounds_horizontally(screen_x) {
       return None;
     }
-    let logic_x = (screen_x - self.screen_view.offset.x) as f32 / self.logical_pixel_width;
-    return Some(logic_x as usize);
+    let logic_x = (screen_x as i64 - self.screen_view.offset.x) as f32 / self.logical_pixel_width;
+    return Some(logic_x as i64);
   }
 
-  pub fn get_logical_position_y(&self, screen_y: usize) -> Option<usize> {
+  pub fn get_logical_position_y(&self, screen_y: usize) -> Option<i64> {
     if !self.in_screen_bounds_vertically(screen_y) {
       return None;
     }
-    let logic_y = (screen_y - self.screen_view.offset.y) as f32 / self.logical_pixel_height;
-    return Some(logic_y as usize);
+    let logic_y = (screen_y as i64 - self.screen_view.offset.y) as f32 / self.logical_pixel_height;
+    return Some(logic_y as i64);
   }
 
   fn in_screen_bounds_horizontally(&self, screen_x: usize) -> bool {
-    (self.screen_view.offset.x..self.screen_view.offset.x + self.screen_view.size.width).contains(&screen_x)
+    (self.screen_view.offset.x..self.screen_view.offset.x + self.screen_view.size.width as i64).contains(&(screen_x as i64))
   }
 
   fn in_screen_bounds_vertically(&self, screen_y: usize) -> bool {
-    (self.screen_view.offset.y..self.screen_view.offset.y + self.screen_view.size.height).contains(&screen_y)
+    (self.screen_view.offset.y..self.screen_view.offset.y + self.screen_view.size.height as i64).contains(&(screen_y as i64))
   }
 
   pub fn get_logical_position(&self, screen_x: usize, screen_y: usize) -> Option<Position> {
